Trying to fix an a shadowing issue with pose

+++ src/lemlib/util.cpp
@@ -70,12 +70,12 @@ float avg(std::vector<float> values) {
  * @brief Return the average of a vector of numbers
  *
  * @param values
- * @return casted float
+ * @return double
  */
 double avg(std::vector<double> values) {
     double sum = 0;
     for (double value : values) { sum += value; }
-    return (float)sum / values.size();
+    return sum / values.size();
 }
 
 /**
